+++ lib/castle/request.rb
@@ -8,24 +8,18 @@ module Castle
       @headers = headers
     end
 
-    def build_query(endpoint)
-      request = Net::HTTP::Get.new(
-        "#{@config.api_endpoint.path}/#{endpoint}", @headers
-      )
-      add_basic_auth(request)
-      request
-    end
-
     def build(endpoint, args, method)
-      request = Net::HTTP.const_get(method.to_s.capitalize).new(
-        "#{@config.api_endpoint.path}/#{endpoint}", @headers
-      )
-      request.body = ::Castle::Utils.replace_invalid_characters(args).to_json
+      request = Net::HTTP.const_get(method.to_s.capitalize).new(build_url(endpoint), @headers)
+      request.body = ::Castle::Utils.replace_invalid_characters(args).to_json unless method == :get
       add_basic_auth(request)
       request
     end
 
     private
+
+    def build_url(endpoint)
+      "/#{@config.url_prefix}/#{endpoint}"
+    end
 
     def add_basic_auth(request)
       request.basic_auth('', @config.api_secret)