+++ spec/spec_helper.rb
@@ -13,7 +13,6 @@ require 'castle'
 
 Castle.configure do |config|
   config.api_secret = 'secret'
-  config.api_endpoint = 'https://api.castle.io/v1'
 end
 
 WebMock.disable_net_connect!(allow_localhost: true)